split: split & remap throttle events
the throttle this is made for (VPC MT-50 Throttle CM) has many
buttons and axis than wine can handle.  So the events are split
over two new virtual devices: grip and base.

For wine to recognise the virutal device it must atleast have:
an X-axis, Y-axis, and two buttons.

The remapped ABS events are configured with fuzz=1 so that the
source values are passed as-is (ie without another step of fuzzing).

Now I can use all the buttons and axis my throttle in Elite:Dangerous.
  o7